AF3_3 - Fixing reorganized file imports
Without fixing all relevant import paths it was impossible to continue development, thus it had to be done.
